+++ automation_tasks_rs/src/main.rs
@@ -162,11 +162,13 @@ fn task_build() {
     println!(
         r#"
   {YELLOW}After `cargo auto build`, run the compiled binary, examples and/or tests{RESET}
-{GREEN}./target/debug/{package_name} print world{RESET}
+{GREEN}alias {package_name}=target/debug/{package_name}{RESET}
+
+{GREEN}{package_name} print world{RESET}
   {YELLOW}If ok then{RESET}
-{GREEN}./target/debug/{package_name} upper world{RESET}
+{GREEN}{package_name} upper world{RESET}
   {YELLOW}If ok then{RESET}
-{GREEN}./target/debug/{package_name} upper WORLD{RESET}
+{GREEN}{package_name} upper WORLD{RESET}
   {YELLOW}if {package_name} ok then{RESET}
 {GREEN}cargo auto release{RESET}
 "#,
@@ -182,11 +184,13 @@ fn task_release() {
     println!(
         r#"
   {YELLOW}After `cargo auto release`, run the compiled binary, examples and/or tests{RESET}
-{GREEN}./target/release/{package_name} print world{RESET}
+{GREEN}alias {package_name}=target/release/{package_name}{RESET}
+
+{GREEN}{package_name} print world{RESET}
   {YELLOW}If ok then{RESET}
-{GREEN}./target/release/{package_name} upper world{RESET}
+{GREEN}{package_name} upper world{RESET}
   {YELLOW}If ok then{RESET}
-{GREEN}./target/release/{package_name} upper WORLD{RESET}
+{GREEN}{package_name} upper WORLD{RESET}
   {YELLOW}if {package_name} ok then{RESET}
 {GREEN}cargo auto doc{RESET}
 "#,