+++ src/hello_mod.rs
@@ -6,7 +6,7 @@
 
 /// format the hello phrase
 pub fn format_hello_phrase(greet_name: &str) -> String {
-    log::info!("start format_hello_phrase()");
+    tracing::info!("start format_hello_phrase()");
     // return
     format!("Hello {}!", greet_name)
 }
@@ -14,10 +14,11 @@ pub fn format_hello_phrase(greet_name: &str) -> String {
 /// format the hello phrase with uppercase name
 /// if it is already uppercase, return error with thiserror
 pub fn format_upper_hello_phrase(greet_name: &str) -> Result<String, crate::LibraryError> {
-    log::info!("start format_upper_hello_phrase()");
+    tracing::info!("start format_upper_hello_phrase()");
     // shadowing the same variable name:
     let upper_greet_name = make_uppercase(greet_name);
     if upper_greet_name == greet_name {
+        tracing::error!("LibraryError::Uppercase {greet_name}");
         return Err(crate::LibraryError::Uppercase(greet_name.to_string()));
     }