+++ RELEASES.md
@@ -12,6 +12,8 @@ The TODO section is part of the [README.md](https://github.com/automation-tasks-
 
 - return commits
 
+- update
+
 ## Version 0.0.8 (2025-03-29)
 
 - update automation