+++ RELEASES.md
@@ -14,6 +14,8 @@ The TODO section is part of the [README.md](https://github.com/automation-tasks-
 
 - update
 
+- tracing
+
 ## Version 0.0.8 (2025-03-29)
 
 - update automation